modularise widget coordinator code
